Break down database.py into a submodule with seperate files
